requirements.txt
Flask
ddddocr
Pillow
requests
